Properly handle executing an entire namespace of tests.
NUnitLite does not construct a 'Test' object for the namespace
hierarchy. It is a flat list of tests which have fully qualified type
names inside a single container. As such we need to use
a Test.Name.StartsWith instead of checking the parent when
we want to run all tests in the 'Foo.Bar' namespace, such as
Foo.Bar.Baz and Foo.Bar.Bim