Disable the context of CallContext as it doesn't work the way we need
We're more than happy with a single static variable. We don't support
multiple appdomains (or want to support it), and our usage of
async/await does not work with CallContext under .NET 4.5. We end up
getting a null value from CallContext.GetData after an await
completes, which is expected behaviour.